+++ sqlalchemy_hana/dialect.py
@@ -76,9 +76,15 @@ if TYPE_CHECKING:
 if sqlalchemy.__version__ < "2.1":
     from sqlalchemy.sql.ddl import _DropView as BaseDropView
 else:
+    # In SQLAlchemy 2.1, the _DropView class was renamed to DropView.
+    # Also CreateView was added
     # isort moves the comments around leading to issues
     # isort: off
-    # In SQLAlchemy 2.1, the _DropView class was renamed to DropView.
+    # pylint:disable-next=no-name-in-module
+    from sqlalchemy.sql.ddl import (  # type: ignore[attr-defined]
+        CreateView as BaseCreateView,
+    )
+
     # pylint:disable-next=no-name-in-module
     from sqlalchemy.sql.ddl import (  # type: ignore[attr-defined,no-redef]
         DropView as BaseDropView,
@@ -459,6 +465,11 @@ class HANADDLCompiler(compiler.DDLCompiler):
         text += self.define_constraint_deferrability(constraint)
         return text
 
+    def visit_create_table_as(self, element: Any, **kw: Any) -> str:
+        result = super().visit_create_table_as(element, **kw)  # type: ignore[misc]
+        prefix, sep, select_sql = result.partition(" AS ")
+        return f"{prefix}{sep}({select_sql})"
+
     @override
     def visit_create_table(self, create: CreateTable, **kw: Any) -> str:
         table = create.element
@@ -502,9 +513,11 @@ class HANADDLCompiler(compiler.DDLCompiler):
         )
         return f"{clause} ({expression})"
 
-    def visit_create_view(self, create: CreateView, **kw: Any) -> str:
+    def visit_create_view(self, create: CreateView | BaseCreateView, **kw: Any) -> str:
+        name = getattr(create, "name", getattr(create, "table_name", None))
+        assert name, "Missing view name"
         selectable = self.sql_compiler.process(create.selectable, literal_binds=True)
-        return f"CREATE VIEW {create.name} AS {selectable}"
+        return f"CREATE VIEW {name} AS {selectable}"
 
     @override
     def visit_drop_view(self, drop: DropView | BaseDropView, **kw: Any) -> str: